refactor: 트랜젝션 서비스 추상화
